refactor: create ingest modules with `convert_dataset` function for each use case
