--- sample_data/use_cases/new_york_energy/charts.json
@@ -1,1 +1,0 @@
-[]
